move delivery spec to own file
